added follow_state to implement follow/unfollow button

+++ network/views.py
@@ -94,26 +94,46 @@ def register(request):
 
 
 def profile(request, user):
-    user = User.objects.filter(username = user).first()
+    profile_user = User.objects.filter(username = user).first()
+    current_user = request.user
     # to get followers of user get the no. of users who are following the displayed user
-    followers = len(Follow.objects.filter(following = user))
+    followers = len(Follow.objects.filter(following = profile_user))
 
     # TO get following, get no, of users the current user follows
-    following = len(Follow.objects.filter(follower = user))
+    following = len(Follow.objects.filter(follower = profile_user))
 
-    posts = Post.objects.filter(user = user).order_by('-date')
+    posts = Post.objects.filter(user = profile_user).order_by('-date')
 
     paginator = Paginator(posts, 2)
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
 
-    return render(request, 'network/profile.html', {
-        "username": user,
-        "followers": followers,
-        "following": following,
-        "page_obj": page_obj,
-        "posts": posts
-    })
+    print(profile_user, current_user)
+
+    is_following = Follow.objects.filter(follower = current_user, following = profile_user)
+
+    # Check if logged in user follws the profile user
+    if(is_following):
+        follow_state = True
+        return render(request, 'network/profile.html', {
+            "username": profile_user,
+            "followers": followers,
+            "following": following,
+            "page_obj": page_obj,
+            "posts": posts,
+            "follow_state":follow_state
+        })
+    else:
+        follow_state = False
+        return render(request, 'network/profile.html', {
+            "username": profile_user,
+            "followers": followers,
+            "following": following,
+            "page_obj": page_obj,
+            "posts": posts,
+            "follow_state":follow_state
+        })
+
 
 
 # @login_required
@@ -155,9 +175,9 @@ def following(request):
 
 
 @login_required(login_url = 'login')
-def follow(request, user):
+def follow(request, user, curr_user):
     profile_user = User.objects.get(username = user)
-    current_user = request.user
+    current_user = User.objects.get(username = curr_user)
    
 
     print("\n\n",profile_user,"\n\n")
@@ -167,19 +187,22 @@ def follow(request, user):
     # follow_prof_user = Follow(follower = current_user, following = profile_user)
     # follow_prof_user.save()
 
-    following = Follow.objects.filter(follower = current_user, following = profile_user)
+    if(profile_user != current_user):
 
-    # check if not already following
-    if(not following): 
-        # create new follow object
-        follow_obj = Follow.objects.create(follower = current_user, following = profile_user) 
-        
-        # commit to db
-        follow_obj.save() 
+        following = Follow.objects.filter(follower = current_user, following = profile_user)
+        follow_state = False
 
-        print(f"{current_user} follows {profile_user}")
-    else:
-        following.delete()
-        print(f"{current_user} unfollwed {profile_user}")
-    return redirect('profile', user = profile_user)
-    # return HttpResponse("")
+        # check if not already following
+        if(not following): 
+            # create new follow object
+            follow_obj = Follow.objects.create(follower = current_user, following = profile_user) 
+            
+            # commit to db
+            follow_obj.save() 
+            
+            print(f"{current_user} follows {profile_user}")
+        else:
+            following.delete()
+            print(f"{current_user} unfollwed {profile_user}")
+        # return redirect('profile', user = profile_user)
+        return HttpResponse("")
